// mcpbox/src/_services/strapi/__generated__/api/data-contracts.ts
export interface Error {
  data?: object | object[] | null;
  error: {
    status?: number;
    name?: string;
    message?: string;
    details?: object;
  };
}
// ...
export interface UsersPermissionsRole {
  id?: number;
  name?: string;
  description?: string;
  type?: string;
  /** @format date-time */
  createdAt?: string;
  /** @format date-time */
  updatedAt?: string;
}
// ...
export type UsersPermissionsPermissionsTree = Record<
  string,
  {
    /** every controller of the api */
    controllers?: Record<
      string,
      Record<
        string,
        {
          enabled?: boolean;
          policy?: string;
        }
      >
    >;
  }
>;
// ...
export interface PermissionsListData {
  permissions?: UsersPermissionsPermissionsTree;
}

export type PermissionsListError = Error;

export interface RolesListData {
  roles?: (UsersPermissionsRole & {
    nb_users?: number;
  })[];
}

export type RolesListError = Error;

export interface RolesCreateData {
  ok?: "true";
}

export type RolesCreateError = Error;

export interface RolesDetailData {
  role?: UsersPermissionsRole;
}

export type RolesDetailError = Error;

export interface RolesUpdateData {
  ok?: "true";
}

export type RolesUpdateError = Error;

export interface RolesDeleteData {
  ok?: "true";
}

export type RolesDeleteError = Error;